apps/jobs/pyproject.toml
[project]
name = "evanm-jobs"
version = "1.0.0"
description = "Python scripts for scheduled jobs and automation"
authors = [
    {name = "Evan Mullins", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "schedule>=1.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "networkx>=3.0",
    "pyvis>=0.3.0",
    "python-jobspy>=1.1.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "pytest-cov>=4.1.0"
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=jobs --cov-report=term-missing"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".venv"]